[kernels]: collapse flash_attn_fp8_gfx950 to a flat module on shared attention
Add the fp8-only ds_read_b64_tr_b8 primitive to
kernels/attention/dualwave_swp_common.py and point flash_attn_fp8_gfx950 at the
shared helpers. The kernel stays a single flat module (no per-kernel package)
and re-exports dualwave_splitk_workspace_elems for its public API.
